requirements.txt
redis==5.0.1
pymongo==4.6.1
faker==22.0.0
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
